requirements.txt
PyGithub
requests
setuptools
customtkinter
GitPython
python-dotenv
packaging
tkinterdnd2